feat: basic structure of code is implemented such as components, states and routing for a few components is completed.
This commit solves the problem of modularity of the code, dividing the shelf types in there respective category. With this commit we can see the list of categories of shelf.
State of the component has been created and moving from parent to child component. No book static book has been added, so you would see empty shelf. CSS has not been added yet.